+++ core/impl/src/main/java/com/blazebit/persistence/impl/JoinNodeGathererVisitor.java
@@ -20,6 +20,7 @@ import com.blazebit.persistence.parser.expression.PathExpression;
 import com.blazebit.persistence.parser.expression.VisitorAdapter;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -31,7 +32,14 @@ import java.util.Set;
  */
 public class JoinNodeGathererVisitor extends VisitorAdapter {
 
-    private List<JoinNode> joinNodes;
+    private Collection<JoinNode> joinNodes;
+
+    public JoinNodeGathererVisitor() {
+    }
+
+    public JoinNodeGathererVisitor(Collection<JoinNode> joinNodes) {
+        this.joinNodes = joinNodes;
+    }
 
     public Set<JoinNode> collectNonRootJoinNodes(ResolvedExpression[] identifierExpressions) {
         joinNodes = new ArrayList<>();
@@ -41,7 +49,7 @@ public class JoinNodeGathererVisitor extends VisitorAdapter {
         }
 
         Set<JoinNode> nonRootJoinNodeClosure = new HashSet<>();
-        List<JoinNode> nodes = joinNodes;
+        List<JoinNode> nodes = (List<JoinNode>) joinNodes;
         size = nodes.size();
         for (int i = 0; i < size; i++) {
             JoinNode n = nodes.get(i);