+++ core/impl/src/main/java/com/blazebit/persistence/impl/SubqueryInitiatorFactory.java
@@ -74,9 +74,9 @@ public class SubqueryInitiatorFactory {
         // TODO: set operations?
 
         if (inExists) {
-            subqueryBuilder.selectManager.setDefaultSelect(Arrays.asList(new SelectInfo(mainQuery.expressionFactory.createArithmeticExpression("1"))));
+            subqueryBuilder.selectManager.setDefaultSelect(null, Arrays.asList(new SelectInfo(mainQuery.expressionFactory.createArithmeticExpression("1"))));
         } else {
-            subqueryBuilder.selectManager.setDefaultSelect(builder.selectManager.getSelectInfos());
+            subqueryBuilder.selectManager.setDefaultSelect(null, builder.selectManager.getSelectInfos());
         }
 
         if (listener != null) {